excelerate: Create Excel Spreadsheets for Supplementary Tables

The excelerate package simplifies the creation of supplementary tables for research papers. It uses the openxlsx R package to produce tables in a consistent format. You can easily turn data frames into Excel files, with accompanying metadata in a README sheet, and save them as numbered supplementary files.

See also

Author

Maintainer: Amelia Edmondson-Stait aedmond3@ed.ac.uk (ORCID) [copyright holder]

Authors:

Mark James Adams mark.adams@ed.ac.uk (ORCID)